The BeanConstraintService now handles @Embeddable.
For example when a Person has an @Embeddable Contact which in turn
has an @Embeddable Address it will now return constraints for each
field of the embeddable.

The end result is a HashMap<String, PropertyConstraintDescription> that
will look like this:

```
{
  "Person": {
  {
    "name": #PropertyConstraintDescription
    "contact.address.city": #PropertyConstraintDescription
    "contact.address.streetAndName": #PropertyConstraintDescription
  }
}
```

Each embeddable will start a new '.' in the key name.

Also the key used for the @Entity will now use its simpleName instead
of the normal name.

Closes #41.