Fixed SqlDatabasePopulator when datasource is configured with autocommit = false.
The Spring ResourceDatabasePopulator expects the given sql connection to have autocommit = true.
When this connection is retrieved from a datasource that is configured with autocommit = false, the sql statements
from the resource are never committed to the database.

This fix sets the autocommit of the connection to true thus overriding the datasource setting for this property.
After the `ResourceDatabasePopulator.populate(connection)` returns, the original autocommit setting of the connection is put back on it
to respect the datasource setting. This is important when the datasource is configured with a connection pool which reuses connections.